pyproject.toml
[build-system]
requires = ["hatchling>=1.24.0"]
build-backend = "hatchling.build"

[project]
name = "conversation-to-spec"
version = "0.1.0"
description = "Python CLI prototype that converts conversations into structured software requirement drafts."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pydantic>=2.7.0,<3.0.0",
  "pyyaml>=6.0.0,<7.0.0",
  "transformers>=4.40.0,<5.0.0",
  "torch>=2.2.0",
]

[project.scripts]
conversation-to-spec = "app.main:main"

[dependency-groups]
dev = [
  "pytest>=8.0.0,<9.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
default-groups = ["dev"]

[tool.pytest.ini_options]
pythonpath = ["."]
